Day 9 I think
got some of the structure for leveling up good still have to do more work but finally got leveling up to increase ability stats.

Will no be adding an autogun ability which is a turret that attaches to the player and will shoot anything that gets close to the player.